requirements.txt
Django==1.11
django-filter==2.2.0
djangorestframework==3.10.3
pkg-resources==0.0.0
pytz==2019.3
